requirements.txt
pydantic
mlflow
pandas
python-dotenv
openai
openai-agents
PyYAML
strands-agents
strands-agents-tools